Make more obvious when imds mock fails to start
There is a flaky test and I believe it fails because the imds mock fails
to start due to the config file not being found. I'm not sure why that
happens, but this should make more obvious if that is the case by
failing early instead of letting the test continue.