Validate remoteNetworkConfig configured for Hybrid
When reading the cluster details from the EKS API to populate the
cluster config, make sure that remoteNetworkConfig is set if it's
supposed to be a Hybrid Node. This allows to fail fast during init
instead of having to debug the misconfiguration issue when the Control
Plane can't connect to the node.

This required a custom de-serialization for the DescribeCluster API call
since the new field is still not available through the API. When it
does, we can remove that file completely.

I have also moved things around to allow to make the describe cluster
call and validation earlier in the flow. Before, it was being done
during the kubelet configuration step, which is way further down the
flow, almost at the end.

Now instead, we do it during the Enrich config step. Which makes sense
since that's what we are doing, enriching the provided config with
cluster details coming from the API. However, this presents a challenge:
now enrich requires AWS creds. This meant I also had to move
ConfigureAws to earlier in the flow. The plot thicken: SSM requires the
agent to be registered and started before we have creds and this
happened in a different step: PreProcessDaemon. In this case I just made
this logic part of ConfigureAws.

Since there are some "config" defaults that are static and don't require
any external dependency, I moved those to PopoluateNodeConfigDefaults,
so we can ran those early in the flow before even having creds, and we
can maintain the invariants that all downstream code expects.

Update test/e2e/nodeadm_test.go

Co-authored-by: Abhinav Pandey <abhinavmpandey08@gmail.com>