Register with SSM agent instead of setup CLI
This ensures that there is no "checksum" error so the setup-cli and
agent don't need to be re-downloaded during init.

The downside is that we lose some of the functionality provided by the
setup CLI, like verifying if the instance is already registered and
making sure the agent is running. So we have to implement this
ourselves.