Wait until restart ssm agent job completes
Certain systemd operation we run are async, thanks to the dbus library
we use. However, in some cases like SSM, following operations expect the
agent to be up and running. Although we were already waiting for the
service to be running, there was a race condition where the job hadn't
even started before we checked the status, hence continuing with the
init flow. This caused the SSM creds to not be available yet by the time
we needed them.