Wait for AWS credentials to available for SSM
Sometimes there is a delay between the ssm agent being restarted and
reaching running state and the aws creds file to be present on disk.
This tries to avoid the race condition where we read the aws config too
fast.

In addition, we now specifically look for the creds file and avoid
looking for config files or IMDS, forcing the go SDK to either use the
creds file from disk or fail.