Make retry cmd command info level
Debug can be annoying when the error is not retryable, since but default
it won't print, so for commands like install or upgrade, they user has
to sit for 20 minutes until it times out. Instead, let's print the error
on each retry by default, so the user can abort if the error needs
manual intervention. In the future, we can try to catch non retryable
errors and just abort.