Fix make lint target
The GOLANGCI_LINT is defined after the lint target, which means that is
empty if you only run make lint. If you haven't downloaded the binary
before, this won't try to download the binary and will fail.

We could have instead moved the whole development block, but that also
affects the order in which we show the blocks in the help. And I think
development makes more sense at the top since it's more important. The
other option wad to define the GOLANGCI_LINT variable at the top, but
then we lose the nice encapsulation of having it defined in the build
section with the targets that need it.