Capture EC2 serial console output of nodes
This will help with situations where the node doesn't join the cluster
and we have no clue why.

Normally, if the problem is nodeadm init or install, this gets reflected
in the cloud-init output log, which is captured by the logs collector
and uploaded to S3. However, there are times where log collector fails
or we can't even ssh into the node to run it. In these cases, we are
left with no information about what happened.

By capturing the serial output, not only we will have some information
in those otherwise indebuggable scenarios, but we should be able to see
in real time the node init process in the test output as well as any
other useful information printed during boot. In addition to stdout, the
serial console logs are also written to a file in disk. The CI pipeline
will persist this file with the other pipeline artifacts in S3.

Co-authored-by: Jackson West <jgw@amazon.com>
Co-authored-by: Guillermo Gaston <gaslor@amazon.com>