Respect no-color option in e2e test logger
We use the ginkgo no-color option when running tests on CI. This avoids
terminal control characters to appear in the logs. However, this only
controls the logs emitted by ginkgo and not the one we print. As a
result, all the "INFO" log lines showed up mangled in CI logs.