Fix bug on errors.IsType helper
If we take an empty interface, the errors.As receives as such instead of
as the original type passed to our helper. Since all errors can be
assigned to an empty interface, our helper always returns true. By
making the function take a generic type we both ensure that errors.As
receives the right type as well as enforcing that our helper only takes
types that implement error, which is required by errors.As.