Print logs with readable time and disable stack
Although perhaps easier to parse, timestamps in epoch is almos
unreadable. ISO8601 should make easier to debug timing issues.

In addition, I find quite annoying to have the stacktrace on the fatal
print line and the end when the command files because this is captured
where Fatal is called and not where the error originated. Hence, it's
always "main.go.....runtime/proc.go...". This is useless and creates
noise when trying to figure out what the actual error is.